qemu: run tasks through vm and cross roots

+++ tasks/core/combine_overlays/combine.sh
@@ -3,7 +3,12 @@
 ROOTFS="$DS_WORK/rootfs/"
 INSTALL="$DS_WORK/overlays/"
 
+shopt -s nullglob
 for dir in $INSTALL/*/
 do
-    rsync -aKHAX --numeric-ids "$dir" "$ROOTFS"/
+    if [[ -f "$dir/overlay.tar" ]]; then
+        tar --xattrs --acls --numeric-owner -C "$ROOTFS" -xf "$dir/overlay.tar"
+    else
+        rsync -aKHAX --numeric-ids "$dir" "$ROOTFS"/
+    fi
 done
